Add Plate Badge Display Components
Break plate badge display components into two separate componenets.
AddByInput and AddByPlate
